修复定时清除缓存作业的BUG

+++ app/task/clearCacheTask.js
@@ -3,48 +3,57 @@ import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import nedb from '../utils/dbUtil';
+import { cherryCached, extractTracks } from '../api/cache';
 
 const debug = require('debug')('clearCacheTask');
 
-const dbPath = path.join(os.homedir(), '.bgm', 'player.db');
 const cachePath = path.join(os.homedir(), '.bgm');
 
-const deleteFolderRecursive = function(dirPath) {
+async function deleteCache(dirPath) {
   if (fs.existsSync(dirPath)) {
-    fs.readdirSync(dirPath).forEach(file => {
+    const cachedFiles = await getCachedFiles();
+    fs.readdirSync(dirPath).forEach(async file => {
       const curPath = `${dirPath}/${file}`;
       if (fs.statSync(curPath).isDirectory()) {
         // recurse
-        deleteFolderRecursive(curPath);
+        await deleteCache(curPath);
       } else {
         // delete file
+        if (cachedFiles.includes(curPath)) {
+          return cachedFiles.splice(cachedFiles.indexOf(curPath), 1);
+        }
+        if (nedb.dbPath === curPath) {
+          return;
+        }
         fs.unlinkSync(curPath);
       }
     });
-    fs.rmdirSync(dirPath);
+    // fs.rmdirSync(dirPath);
   }
-};
+}
+
+/**
+ * 获取播放计划中的缓存文件列表
+ * @returns {Promise<any[]>}
+ */
+export async function getCachedFiles() {
+  const plan = await nedb.getPlayerPlan();
+  const {
+    cachedPlaylists,
+    cachedScrollAudioMessage,
+    cachedAlarmAudioMessages
+  } = await cherryCached(plan);
+  const cachedFiles = [
+    ...extractTracks(cachedPlaylists),
+    ...cachedAlarmAudioMessages,
+    ...cachedScrollAudioMessage
+  ].map(item => item.filePathName);
+  return cachedFiles;
+}
 
 // @TODO: 现在是全部清空，容易招致BUG，后面可以选择性清空
 export async function clearCache() {
-  try {
-    // 保存激活码
-    const activeCode = await nedb.getActiveCode();
-    // 清除缓存
-    deleteFolderRecursive(cachePath);
-    // 重建缓存
-    fs.mkdirSync(cachePath, err => {
-      if (err) {
-        //
-      }
-    });
-    fs.writeFileSync(dbPath, '', { encoding: 'utf-8' });
-    await nedb.insert({ activeCode });
-    return true;
-  } catch (e) {
-    debug(e.toString());
-    return false;
-  }
+  await deleteCache(cachePath);
 }
 
 /**
